update corrdiff latlon grid handling and add tests
